pyproject.toml
[tool.poetry]
name = "taskmaster"
version = "0.1.0"
description = ""
authors = ["Maksuel Nascimento <mk-nascimento@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.3"
alembic = "^1.13.1"
mariadb = "^1.1.10"
sqlalchemy = "^2.0.30"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.7"
httpx = "^0.27.0"
pre-commit = "^3.7.1"
factory-boy = "^3.3.0"

[tool.ruff]
line-length = 88
extend-exclude = ["alembic"]

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "W", "PL", "PT"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
post_test = "coverage html"
test = "pytest -s -x --cov=taskmaster -vv"
dev = "fastapi dev taskmaster/main.py"
req = 'poetry export --without-hashes -f requirements.txt -o requirements.txt'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
